Separate Inventory as 2 subclasses
